% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bazin.R
\name{bazin}
\alias{bazin}
\title{O preço máximo da ação pelo método de Décio Bazin}
\usage{
bazin(proventos)
}
\arguments{
\item{proventos}{Soma de todos os proventos (dividendos e juros sobre capital próprio) 
entregues nos últimos doze meses.}
}
\value{
O preço máximo da ação, para que ela renda 6\% ao ano na forma de proventos.
}
\description{
Método de Décio Bazin, para encontrar o preço máximo de uma ação, 
      para que ela renda 6\% ao ano, baseado nos proventos distribuídos nos últimos 
      12 meses.
}
\examples{

Proventos distribuídos pela Itaúsa no ano de 2019

bazin(1.18)


}
\author{
Guilherme Hathy
}
