% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roa.R
\name{roa}
\alias{roa}
\title{ROA - Return on Assets}
\usage{
roa(lucroLiquido, ativoTotal)
}
\arguments{
\item{lucroLiquido}{Lucro Líquido do período.}

\item{ativoTotal}{Ativo Total do período. \deqn{Ativo Circulante + Ativo Não Circulante}}
}
\value{
O ROA do período.
}
\description{
ROA (Return on Assets) ou Retorno sobre o Ativo é um indicador que apresenta 
     como a empresa é rentável em relação ao seu total de ativos, o ROA fornece uma visão 
     de quão eficiente é a gestão da empresa na utilização de seus ativos para gerar ganhos.
     
     \deqn{ROA = \frac{Lucro\ Líquido}{Ativo\ Total}}
}
\examples{

Dados de  Itaúsa no ano de 2019

lucroLiquido <- 10569
ativoTotal <- 67476

roa(lucroLiquido, ativoTotal)


}
\author{
Guilherme Hathy
}
