% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roe.R
\name{roe}
\alias{roe}
\title{ROE - Return On Equity}
\usage{
roe(lucroLiquido, patrimonioLiquido)
}
\arguments{
\item{lucroLiquido}{Lucro Líquido do período.}

\item{patrimonioLiquido}{Patrimônio Líquido do período.}
}
\value{
O ROE do período.
}
\description{
ROE (Return On Equity) ou Retorno Sobre o Patrimônio Líquido é uma porcentagem 
     usada para medir o quão eficiente uma empresa é para a capacidade de geração de lucros 
     utilizando o capital investido pelos acionistas.
     
     \deqn{ROE = \frac{Lucro\ Líquido}{Patrimônio\ Líquido}}
}
\examples{

Dados de  Itaúsa no ano de 2019

lucroLiquido <- 10569
patrimonioLiquido <- 58357

roe(lucroLiquido, patrimonioLiquido)


}
\author{
Guilherme Hathy
}
